+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fix when 'Renew Session' button is shown #460
 - Ensure app works on the iPad
 - Ensure Settings and Help buttons are visible on the iPad